* New 2.x scripts
* New games
* Games updates

+++ play.it-2/games/play-convoy.sh
@@ -1,8 +1,9 @@
-#!/bin/sh -e
+#!/bin/sh
 set -o errexit
 
 ###
 # Copyright (c) 2015-2019, Antoine "vv221/vv222" Le Gonidec
+# Copyright (c) 2016-2019, Solène "Mopi" Huault
 # All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without
@@ -30,48 +31,64 @@ set -o errexit
 
 ###
 # Convoy
-# build native Linux packages from the original installers
+# build native packages from the original installers
 # send your bug reports to vv221@dotslashplay.it
 ###
 
-script_version=20180224.1
+script_version=20190611.3
 
 # Set game-specific variables
 
 GAME_ID='convoy'
 GAME_NAME='Convoy'
 
-ARCHIVES_LIST='ARCHIVE_GOG'
-
-ARCHIVE_GOG='gog_convoy_2.4.0.7.sh'
+ARCHIVE_GOG='convoy_1_1_54_27852.sh'
 ARCHIVE_GOG_URL='https://www.gog.com/game/convoy'
-ARCHIVE_GOG_MD5='2d66599173990eb202a43dbc547c80f5'
+ARCHIVE_GOG_MD5='2f7dd6c597e07638650cc01883e0367f'
 ARCHIVE_GOG_SIZE='860000'
-ARCHIVE_GOG_VERSION='1.1.51-gog2.4.0.7'
+ARCHIVE_GOG_VERSION='1.1.54-gog27852'
+ARCHIVE_GOG_TYPE='mojosetup'
+
+ARCHIVE_GOG_OLD2='convoy_1_1_53_27205.sh'
+ARCHIVE_GOG_OLD2_MD5='cda02a99f12adc608a0193f75fc9d7d3'
+ARCHIVE_GOG_OLD2_SIZE='860000'
+ARCHIVE_GOG_OLD2_VERSION='1.1.53-gog27205'
+ARCHIVE_GOG_OLD2_TYPE='mojosetup'
+
+ARCHIVE_GOG_OLD1='convoy_1_1_52_26363.sh'
+ARCHIVE_GOG_OLD1_MD5='99b331906d75443f08c4f787bc83a7ef'
+ARCHIVE_GOG_OLD1_SIZE='860000'
+ARCHIVE_GOG_OLD1_VERSION='1.1.52-gog26363'
+ARCHIVE_GOG_OLD1_TYPE='mojosetup'
+
+ARCHIVE_GOG_OLD0='gog_convoy_2.4.0.7.sh'
+ARCHIVE_GOG_OLD0_MD5='2d66599173990eb202a43dbc547c80f5'
+ARCHIVE_GOG_OLD0_SIZE='860000'
+ARCHIVE_GOG_OLD0_VERSION='1.1.51-gog2.4.0.7'
 
 ARCHIVE_DOC_DATA_PATH='data/noarch/docs'
-ARCHIVE_DOC_DATA_FILES='./*'
+ARCHIVE_DOC_DATA_FILES='*'
 
 ARCHIVE_GAME_BIN32_PATH='data/noarch/game'
-ARCHIVE_GAME_BIN32_FILES='./*.x86 ./*_Data/*/x86'
+ARCHIVE_GAME_BIN32_FILES='Convoy.x86 Convoy_Data/*/x86'
 
 ARCHIVE_GAME_BIN64_PATH='data/noarch/game'
-ARCHIVE_GAME_BIN64_FILES='./*.x86_64 ./*_Data/*/x86_64'
+ARCHIVE_GAME_BIN64_FILES='Convoy.x86_64 Convoy_Data/*/x86_64'
 
 ARCHIVE_GAME_DATA_PATH='data/noarch/game'
-ARCHIVE_GAME_DATA_FILES='./*_Data/level* ./*_Data/PlayerConnectionConfigFile ./*_Data/*.assets ./*_Data/mainData ./*_Data/Mono/etc ./*_Data/Managed ./*_Data/Resources ./*_Data/ScreenSelector.png ./*_Data/*.resS ./*_Data/StreamingAssets ./Modding_Tools'
+ARCHIVE_GAME_DATA_FILES='Convoy_Data'
 
 CONFIG_FILES='./*.ini'
 DATA_DIRS='./logs'
 
 APP_MAIN_TYPE='native'
+# shellcheck disable=SC2016
+APP_MAIN_PRERUN='export TERM="${TERM%-256color}"'
 APP_MAIN_EXE_BIN32='Convoy.x86'
 APP_MAIN_EXE_BIN64='Convoy.x86_64'
 # shellcheck disable=SC2016
 APP_MAIN_OPTIONS='-logFile ./logs/$(date +%F-%R).log'
-APP_MAIN_ICONS_LIST='APP_MAIN_ICON'
-APP_MAIN_ICON='*_Data/Resources/UnityPlayer.png'
-APP_MAIN_ICON_RES='128'
+APP_MAIN_ICON='Convoy_Data/Resources/UnityPlayer.png'
 
 PACKAGES_LIST='PKG_BIN32 PKG_BIN64 PKG_DATA'
 
@@ -79,50 +96,71 @@ PKG_DATA_ID="${GAME_ID}-data"
 PKG_DATA_DESCRIPTION='data'
 
 PKG_BIN32_ARCH='32'
-PKG_BIN32_DEPS="$PKG_DATA_ID glibc libstdc++ glu xcursor"
+PKG_BIN32_DEPS="$PKG_DATA_ID glibc libstdc++ gtk2 glx"
+PKG_BIN32_DEPS_ARCH='lib32-gdk-pixbuf2 lib32-glib2'
+PKG_BIN32_DEPS_DEB='libgdk-pixbuf2.0-0, libglib2.0-0'
+PKG_BIN32_DEPS_GENTOO='x11-libs/gdk-pixbuf[abi_x86_32] dev-libs/glib[abi_x86_32]'
+# Keep compatibility with old archives
+PKG_BIN32_DEPS_GOG_OLD0="$PKG_DATA_ID glibc libstdc++ glu glx xcursor"
+PKG_BIN32_DEPS_ARCH_GOG_OLD0='lib32-libx11'
+PKG_BIN32_DEPS_DEB_GOG_OLD0='libx11-6'
+PKG_BIN32_DEPS_GENTOO_GOG_OLD0='x11-libs/libX11[abi_x86_32]'
 
 PKG_BIN64_ARCH='64'
 PKG_BIN64_DEPS="$PKG_BIN32_DEPS"
+PKG_BIN64_DEPS_ARCH='gdk-pixbuf2 glib2'
+PKG_BIN64_DEPS_DEB="$PKG_BIN32_DEPS_DEB"
+PKG_BIN64_DEPS_GENTOO='x11-libs/gdk-pixbuf dev-libs/glib'
+# Keep compatibility with old archives
+PKG_BIN64_DEPS_GOG_OLD0="$PKG_BIN32_DEPS_GOG_OLD0"
+PKG_BIN64_DEPS_ARCH_GOG_OLD0='libx11'
+PKG_BIN64_DEPS_DEB_GOG_OLD0="$PKG_BIN32_DEPS_DEB_GOG_OLD0"
+PKG_BIN64_DEPS_GENTOO_GOG_OLD0='x11-libs/libX11'
 
 # Load common functions
 
-target_version='2.5'
+target_version='2.11'
 
 if [ -z "$PLAYIT_LIB2" ]; then
-	[ -n "$XDG_DATA_HOME" ] || XDG_DATA_HOME="$HOME/.local/share"
-	if [ -e "$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh" ]; then
-		PLAYIT_LIB2="$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh"
-	elif [ -e './libplayit2.sh' ]; then
-		PLAYIT_LIB2='./libplayit2.sh'
-	else
-		printf '\n\033[1;31mError:\033[0m\n'
-		printf 'libplayit2.sh not found.\n'
-		exit 1
-	fi
+	: "${XDG_DATA_HOME:="$HOME/.local/share"}"
+	for path in\
+		"$PWD"\
+		"$XDG_DATA_HOME/play.it"\
+		'/usr/local/share/games/play.it'\
+		'/usr/local/share/play.it'\
+		'/usr/share/games/play.it'\
+		'/usr/share/play.it'
+	do
+		if [ -e "$path/libplayit2.sh" ]; then
+			PLAYIT_LIB2="$path/libplayit2.sh"
+			break
+		fi
+	done
 fi
-#shellcheck source=play.it-2/lib/libplayit2.sh
+if [ -z "$PLAYIT_LIB2" ]; then
+	printf '\n\033[1;31mError:\033[0m\n'
+	printf 'libplayit2.sh not found.\n'
+	exit 1
+fi
+# shellcheck source=play.it-2/lib/libplayit2.sh
 . "$PLAYIT_LIB2"
 
 # Extract game data
 
 extract_data_from "$SOURCE_ARCHIVE"
-
-for PKG in $PACKAGES_LIST; do
-	organize_data "DOC_${PKG#PKG_}"  "$PATH_DOC"
-	organize_data "GAME_${PKG#PKG_}" "$PATH_GAME"
-done
-
+prepare_package_layout
 rm --recursive "$PLAYIT_WORKDIR/gamedata"
 
 # Write launchers
 
 for PKG in 'PKG_BIN32' 'PKG_BIN64'; do
-	write_launcher 'APP_MAIN'
+	launchers_write 'APP_MAIN'
 done
 
 # Build package
 
-postinst_icons_linking 'APP_MAIN'
+PKG='PKG_DATA'
+icons_linking_postinst 'APP_MAIN'
 write_metadata 'PKG_DATA'
 write_metadata 'PKG_BIN32' 'PKG_BIN64'
 build_pkg
@@ -133,10 +171,6 @@ rm --recursive "$PLAYIT_WORKDIR"
 
 # Print instructions
 
-printf '\n'
-printf '32-bit:'
-print_instructions 'PKG_DATA' 'PKG_BIN32'
-printf '64-bit:'
-print_instructions 'PKG_DATA' 'PKG_BIN64'
+print_instructions
 
 exit 0
