+++ play.it-2/games/play-system-shock-2.sh
@@ -1,4 +1,4 @@
-#!/bin/sh -e
+#!/bin/sh
 set -o errexit
 
 ###
@@ -30,56 +30,82 @@ set -o errexit
 
 ###
 # System Shock 2
-# build native Linux packages from the original installers
+# build native packages from the original installers
 # send your bug reports to vv221@dotslashplay.it
 ###
 
-script_version=20180413.1
+script_version=20190520.1
 
 # Set game-specific variables
 
 GAME_ID='system-shock-2'
 GAME_NAME='System Shock 2'
 
-ARCHIVES_LIST='ARCHIVE_GOG ARCHIVE_GOG_OLD ARCHIVE_GOG_OLDER'
+ARCHIVE_GOG='setup_system_shock_2_2.47_nd_(22087).exe'
+ARCHIVE_GOG_URL='https://www.gog.com/game/system_shock_2'
+ARCHIVE_GOG_MD5='cc2ff390b566364447dc5bd05757fe57'
+ARCHIVE_GOG_SIZE='670000'
+ARCHIVE_GOG_VERSION='2.47-gog22087'
+ARCHIVE_GOG_TYPE='innosetup1.7'
 
-ARCHIVE_GOG='setup_system_shock_2_2.46_update_2_(18733).exe'
-ARCHIVE_GOG_MD5='39fab64451ace95966988bb90c7bb17e'
-ARCHIVE_GOG_SIZE='680000'
-ARCHIVE_GOG_VERSION='2.46.2-gog18733'
+ARCHIVE_GOG_OLD3='setup_system_shock_2_2.46_update_3_(19935).exe'
+ARCHIVE_GOG_OLD3_MD5='cdafcdea01556eccab899f94503843df'
+ARCHIVE_GOG_OLD3_SIZE='670000'
+ARCHIVE_GOG_OLD3_VERSION='2.46.3-gog19935'
+ARCHIVE_GOG_OLD3_TYPE='innosetup1.7'
 
-ARCHIVE_GOG_OLD='setup_system_shock_2_2.46_update_(18248).exe'
-ARCHIVE_GOG_OLD_MD5='b76803e4a632b58527eada8993999143'
-ARCHIVE_GOG_OLD_SIZE='690000'
-ARCHIVE_GOG_OLD_VERSION='2.46.1-gog18248'
+ARCHIVE_GOG_OLD2='setup_system_shock_2_2.46_update_2_(18733).exe'
+ARCHIVE_GOG_OLD2_MD5='39fab64451ace95966988bb90c7bb17e'
+ARCHIVE_GOG_OLD2_SIZE='680000'
+ARCHIVE_GOG_OLD2_VERSION='2.46.2-gog18733'
 
-ARCHIVE_GOG_OLDER='setup_system_shock_2_2.46_nd_(11004).exe'
-ARCHIVE_GOG_OLDER_MD5='98c3d01d53bb2b0dc25d7ed7093a67d3'
-ARCHIVE_GOG_OLDER_SIZE='680000'
-ARCHIVE_GOG_OLDER_VERSION='2.46-gog11004'
+ARCHIVE_GOG_OLD1='setup_system_shock_2_2.46_update_(18248).exe'
+ARCHIVE_GOG_OLD1_MD5='b76803e4a632b58527eada8993999143'
+ARCHIVE_GOG_OLD1_SIZE='690000'
+ARCHIVE_GOG_OLD1_VERSION='2.46.1-gog18248'
 
-ARCHIVE_DOC_DATA_PATH='app'
-ARCHIVE_DOC_DATA_FILES='./*.pdf ./*.txt ./*.wri ./doc ./editor/*.txt'
+ARCHIVE_GOG_OLD0='setup_system_shock_2_2.46_nd_(11004).exe'
+ARCHIVE_GOG_OLD0_MD5='98c3d01d53bb2b0dc25d7ed7093a67d3'
+ARCHIVE_GOG_OLD0_SIZE='680000'
+ARCHIVE_GOG_OLD0_VERSION='2.46-gog11004'
 
-ARCHIVE_GAME1_BIN_PATH='app'
-ARCHIVE_GAME1_BIN_FILES='./*.ax ./*.bnd ./*.cfg ./*.exe ./*.osm ./7z.dll ./d3dx9_43.dll ./ffmpeg.dll ./fmsel.dll ./ir41_32.dll ./ir50_32.dll ./lgvid.dll ./msvcrt40.dll ./editor/*.cfg ./editor/*.dll ./editor/*.exe ./microsoft.vc90.crt'
+ARCHIVE_DOC_DATA_PATH='.'
+ARCHIVE_DOC_DATA_FILES='*.pdf *.txt *.wri doc editor/*.txt'
+# Keep compatibility with old archives
+ARCHIVE_DOC_DATA_PATH_GOG_OLD1='app'
+ARCHIVE_DOC_DATA_PATH_GOG_OLD2='app'
+ARCHIVE_DOC_DATA_PATH_GOG_OLD0='app'
 
-ARCHIVE_GAME2_BIN_PATH='app/__support/app'
-ARCHIVE_GAME2_BIN_FILES='./*.cfg ./*.ini'
+ARCHIVE_GAME0_BIN_PATH='.'
+ARCHIVE_GAME0_BIN_FILES='*.ax *.bnd *.cfg *.exe *.osm 7z.dll d3dx9_43.dll ffmpeg.dll fmsel.dll ir41_32.dll ir50_32.dll lgvid.dll msvcrt40.dll editor/*.cfg editor/*.dll editor/*.exe microsoft.vc90.crt'
+# Keep compatibility with old archives
+ARCHIVE_GAME0_BIN_PATH_GOG_OLD1='app'
+ARCHIVE_GAME0_BIN_PATH_GOG_OLD2='app'
+ARCHIVE_GAME0_BIN_PATH_GOG_OLD0='app'
 
-ARCHIVE_GAME_DATA_PATH='app'
-ARCHIVE_GAME_DATA_FILES='./*.bin ./*.dif ./*.dml ./ilist.* ./patch* ./binds ./data ./sq_scripts'
+ARCHIVE_GAME1_BIN_PATH='__support/app'
+ARCHIVE_GAME1_BIN_FILES='*.cfg *.ini'
+# Keep compatibility with old archives
+ARCHIVE_GAME1_BIN_PATH_GOG_OLD1='app/__support/app'
+ARCHIVE_GAME1_BIN_PATH_GOG_OLD2='app/__support/app'
+ARCHIVE_GAME1_BIN_PATH_GOG_OLD0='app/__support/app'
+
+ARCHIVE_GAME_DATA_PATH='.'
+ARCHIVE_GAME_DATA_FILES='*.bin *.dif *.dml ilist.* patch* binds data sq_scripts'
+# Keep compatibility with old archives
+ARCHIVE_GAME_DATA_PATH_GOG_OLD1='app'
+ARCHIVE_GAME_DATA_PATH_GOG_OLD2='app'
+ARCHIVE_GAME_DATA_PATH_GOG_OLD0='app'
 
 CONFIG_FILES='./*.bnd ./*.cfg ./*.ini'
 DATA_DIRS='./current ./save_0 ./save_1 ./save_2 ./save_3 ./save_4 ./save_5 ./save_6 ./save_7 ./save_8 ./save_9 ./save_10 ./save_11 ./save_12 ./save_13 ./save_14'
 DATA_FILES='./*.log'
 
-APP_WINETRICKS="vd=\$(xrandr|grep '\\*'|awk '{print \$1}')"
+APP_WINETRICKS="vd=\$(xrandr|awk '/\\*/ {print \$1}')"
 
 APP_MAIN_TYPE='wine'
 APP_MAIN_EXE='shock2.exe'
 APP_MAIN_ICON='shock2.exe'
-APP_MAIN_ICON_RES='16 32 48 64'
 
 PACKAGES_LIST='PKG_DATA PKG_BIN'
 
@@ -91,44 +117,48 @@ PKG_BIN_DEPS="$PKG_DATA_ID wine winetricks xrandr"
 
 # Load common functions
 
-target_version='2.5'
+target_version='2.11'
 
 if [ -z "$PLAYIT_LIB2" ]; then
-	[ -n "$XDG_DATA_HOME" ] || XDG_DATA_HOME="$HOME/.local/share"
-	if [ -e "$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh" ]; then
-		PLAYIT_LIB2="$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh"
-	elif [ -e './libplayit2.sh' ]; then
-		PLAYIT_LIB2='./libplayit2.sh'
-	else
-		printf '\n\033[1;31mError:\033[0m\n'
-		printf 'libplayit2.sh not found.\n'
-		exit 1
-	fi
+	: "${XDG_DATA_HOME:="$HOME/.local/share"}"
+	for path in\
+		"$PWD"\
+		"$XDG_DATA_HOME/play.it"\
+		'/usr/local/share/games/play.it'\
+		'/usr/local/share/play.it'\
+		'/usr/share/games/play.it'\
+		'/usr/share/play.it'
+	do
+		if [ -e "$path/libplayit2.sh" ]; then
+			PLAYIT_LIB2="$path/libplayit2.sh"
+			break
+		fi
+	done
 fi
-#shellcheck source=play.it-2/lib/libplayit2.sh
+if [ -z "$PLAYIT_LIB2" ]; then
+	printf '\n\033[1;31mError:\033[0m\n'
+	printf 'libplayit2.sh not found.\n'
+	exit 1
+fi
+# shellcheck source=play.it-2/lib/libplayit2.sh
 . "$PLAYIT_LIB2"
 
 # Extract game data
 
 extract_data_from "$SOURCE_ARCHIVE"
+prepare_package_layout
+rm --recursive "$PLAYIT_WORKDIR/gamedata"
 
-for PKG in $PACKAGES_LIST; do
-	organize_data "DOC_${PKG#PKG_}"   "$PATH_DOC"
-	organize_data "GAME_${PKG#PKG_}"  "$PATH_GAME"
-	organize_data "GAME1_${PKG#PKG_}" "$PATH_GAME"
-	organize_data "GAME2_${PKG#PKG_}" "$PATH_GAME"
-done
+# Extract icons
 
 PKG='PKG_BIN'
-extract_and_sort_icons_from 'APP_MAIN'
-move_icons_to 'PKG_DATA'
-
-rm --recursive "$PLAYIT_WORKDIR/gamedata"
+icons_get_from_package 'APP_MAIN'
+icons_move_to 'PKG_DATA'
 
 # Write launchers
 
 PKG='PKG_BIN'
-write_launcher 'APP_MAIN'
+launchers_write 'APP_MAIN'
 
 # Build package
